+++ libs/newsletter-workflow/src/index.ts
@@ -1,2 +1,2 @@
-export * from './lib/newsletter-workflow';
+export * from './lib/steps';
 export * from './lib/schemas';
